frameworks-comparison/resonate_worker/requirements.txt
fastapi
uvicorn
resonate-sdk
